requirements.txt
gradio
pillow
numpy
matplotlib
piexif
opencv-python